refactor:Separating statement parse func
