+++ src/sql/src/parsers.rs
@@ -18,5 +18,6 @@ pub(crate) mod create_parser;
 pub(crate) mod delete_parser;
 pub(crate) mod insert_parser;
 pub(crate) mod query_parser;
+pub(crate) mod show_parser;
 pub(crate) mod tql_parser;
 pub(crate) mod truncate_parser;
